bot/requirements.txt
python-telegram-bot==20.8
psycopg2-binary==2.9.9
python-dotenv==1.0.0
APScheduler==3.10.4
pandas==2.1.4
openpyxl==3.1.2
sqlalchemy==2.0.25
pytz==2023.3
xlsxwriter==3.1.9